refactor(ui): extract feed/pagination.ts — pure page helpers + page-size constants
